limit amount of status percentage that is output
only output percentage every 5 seconds or when it finishes, this avoids
the issue of lots of stdout being written.

Additionally display failed commands in red, and avoid showing too much
of context canceled errors which are a side effect of the main failure
(and can confuse users).
